+++ cmd/frontend/internal/auth/openidconnect/cache.go
@@ -5,7 +5,6 @@ import (
 	"sync"
 	"time"
 
-	oidc "github.com/coreos/go-oidc"
 	"github.com/pkg/errors"
 )
 
@@ -21,7 +20,7 @@ type providerCache struct {
 
 type providerCacheEntry struct {
 	once    sync.Once
-	val     *oidc.Provider
+	val     *provider
 	err     error
 	expires time.Time
 }
@@ -30,7 +29,7 @@ type providerCacheEntry struct {
 // returns it from the cache; otherwise it performs a network request to look up the provider. At
 // most one network request will be in flight for a given issuerURL; later requests block on the
 // original request.
-func (c *providerCache) get(issuerURL string) (*oidc.Provider, error) {
+func (c *providerCache) get(issuerURL string) (*provider, error) {
 	c.mu.Lock()
 	if c.data == nil {
 		c.data = map[string]*providerCacheEntry{}
@@ -44,7 +43,7 @@ func (c *providerCache) get(issuerURL string) (*oidc.Provider, error) {
 
 	fetched := false // whether it was fetched in *this* func call
 	e.once.Do(func() {
-		e.val, e.err = oidc.NewProvider(context.Background(), issuerURL)
+		e.val, e.err = newProvider(context.Background(), issuerURL)
 		e.err = errors.WithMessage(e.err, "retrieving OpenID Connect metadata from issuer")
 		fetched = true